Place the --output-mode in front of job scritp (#28)
Fixes #27

`--output-mode` option was wrongly put in the end of the submit command
which cause the problem that stdout of submit command is not json
parsable. The test was not able to catch it because the command was
defined in the test but not directly get from the scheduler. Test is
update to using the commond generated from scheduler and therefore being
tested directly.