Change `id` field passed to JobInfo from int to str (#30)
fixes #29

The id field of JobInfo is expecting a str. In #24 when parsing JSON
output of `hq job list` the json loads will use the int for the id
parsed directly. Wrong type causes the subtle issue that when job is
waiting it not get into QUEUED state, but immediatly finished and get
nothing to parse from output.