⬆️ Provide support for `aiida-core~=2.7` via `BashCliScheduler`
The following commit in `aiida-core`

https://github.com/aiidateam/aiida-core/commit/954cbdd3ee5127d6618db9d144508505e41cffcc

was released in `aiida-core` v2.7, and breaks existing scheduler plugins. As the commit
message describes, the solution is to change the parent class of the
`HyperQueueScheduler` from `Scheduler` to `BashCliScheduler`. This is done here.

Since the `BashCliScheduler` is only released from `aiida-core` v2.7, we also bump the
dependency to `~=2.7`.

Co-authored-by: Ali Khosravi <khsrali@gmail.com>